switched to tc test framework for api tests
